refactor @polar/core#createMap to be more readable
Relates: #50
